fix client: add gen_proto&parse_proto
